+++ SDDP-iFit-Team3/Model/Goal.swift
@@ -14,13 +14,13 @@ class Goal: NSObject {
     var activityName: String;
     var date : String;
     var duration: Int; // day
-    var progressPercent:Int;
+    var progressPercent: Double;
     var totalExerciseAmount: Int;    // A number for the user to reach their activity target, 50 push.
     var status: String;
     //  var pointsReward: Int;
 
     var userId: String?;
-    init(goalTitle:String,activityName:String,date:String,duration:Int,progressPercent:Int,totalExerciseAmount:Int,status:String) {
+    init(goalTitle:String,activityName:String,date:String,duration:Int,progressPercent:Double,totalExerciseAmount:Int,status:String) {
         self.activityName = activityName;
         self.goalTitle = goalTitle;
         self.date = date;
